+++ src/Home.test.js
@@ -35,16 +35,12 @@ jest.mock('audiomotion-analyzer', () =>
   }))
 );
 
-// Mock react-hint factory
-jest.mock('react-hint', () => {
+// Mock react-tooltip
+jest.mock('react-tooltip', () => {
   const React = require('react');
-  return () =>
-    class MockReactHint extends React.Component {
-      toggleHint = jest.fn();
-      render() {
-        return React.createElement('div', { 'data-testid': 'react-hint' });
-      }
-    };
+  return {
+    Tooltip: (props) => React.createElement('div', { 'data-testid': 'react-tooltip' }),
+  };
 });
 
 const defaultProps = {